apps/ai-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic-settings==2.5.0
openai==1.50.0
python-dotenv==1.0.1
httpx==0.27.0
pymupdf==1.24.11
